omr_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
opencv-python==4.10.0.84
numpy==1.26.4
pdf2image==1.17.0
Pillow==10.4.0
